Package refactoring in examples
