--- examples/src/com/google/inject/examples/Contact.java
@@ -1,3 +1,0 @@
-package com.google.inject.examples;
-
-public class Contact {}
