+++ examples/src/com/google/inject/examples/Contacts.java
@@ -1,5 +1,7 @@
 package com.google.inject.examples;
 
+import com.google.inject.examples.model.Contact;
+
 public interface Contacts {
   Iterable<Contact> findByName(String name);
 }
