+++ examples/src/com/google/inject/examples/memory/FlashMemory.java
@@ -1,4 +1,7 @@
-package com.google.inject.examples;
+package com.google.inject.examples.memory;
+
+import com.google.inject.examples.Contacts;
+import com.google.inject.examples.model.Contact;
 
 import java.util.Collections;
