+++ examples/src/com/google/inject/examples/memory/FromFlash.java
@@ -1,8 +1,8 @@
-package com.google.inject.examples;
+package com.google.inject.examples.memory;
 
 import static java.lang.annotation.RetentionPolicy.*;
 
 import java.lang.annotation.Retention;
 
 @Retention(RUNTIME)
-public @interface FromSim {}
+public @interface FromFlash {}
