+++ examples/src/com/google/inject/examples/model/Contact.java
@@ -1,0 +1,3 @@
+package com.google.inject.examples.model;
+
+public class Contact {}
