+++ examples/src/com/google/inject/examples/model/Phone.java
@@ -1,4 +1,6 @@
-package com.google.inject.examples;
+package com.google.inject.examples.model;
+
+import com.google.inject.examples.Contacts;
 
 public class Phone {
